:recycle: extract to not `.d.ts` files
